requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
APScheduler==3.6.3
asgiref==3.8.1
blinker==1.7.0
cachetools==5.5.0
certifi==2020.12.5
chardet==4.0.0
click==8.1.7
colorama==0.4.6
coloredlogs==15.0
dnspython==2.7.0
email_validator==2.2.0
fastapi==0.115.3
fastapi-cli==0.0.5
Flask==3.0.2
google-ai-generativelanguage==0.6.10
google-api-core==2.21.0
google-api-python-client==2.149.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.3
googleapis-common-protos==1.65.0
greenlet==3.0.3
grpcio==1.67.0
grpcio-status==1.67.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.6
httplib2==0.22.0
httptools==0.6.4
httpx==0.27.2
humanfriendly==10.0
Hypercorn==0.17.3
hyperframe==6.0.1
idna==2.10
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.3
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
packaging==24.1
pluggy==1.5.0
priority==2.0.0
proto-plus==1.25.0
protobuf==5.28.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
pyparsing==3.2.0
pyreadline3==3.4.1
pytest==8.3.3
python-dotenv==1.0.1
python-multipart==0.0.16
python-telegram-bot==13.5
pytz==2021.1
PyYAML==6.0.2
requests==2.25.1
rich==13.9.3
rsa==4.9
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
SQLAlchemy==1.4.14
starlette==0.41.2
toml==0.10.2
tornado==6.1
tqdm==4.66.5
typer==0.12.5
typing_extensions==4.12.2
tzlocal==2.1
uritemplate==4.1.1
urllib3==1.26.5
uvicorn==0.32.0
watchfiles==0.24.0
websockets==13.1
Werkzeug==3.0.2
wsproto==1.2.0
openai